+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/validation/certifiedProduct/InpatientModular2014Validator.java
@@ -89,11 +89,6 @@ public class InpatientModular2014Validator extends CertifiedProduct2014Validator
 					for(PendingCertificationResultTestToolDTO testTool : cert.getTestTools()) {
 						if(testTool.getTestToolId() == null) {
 							product.getErrorMessages().add("There was no test tool found matching '" + testTool.getName() + "' for certification " + cert.getNumber() + ".");
-						} else {
-							TestToolDTO tt = ttDao.getById(testTool.getTestToolId());
-							if(tt != null && tt.isRetired()) {
-								product.getErrorMessages().add("Cannot use test tool '" + tt.getName() + "' since it is retired. Please choose a different test tool for " + cert.getNumber() + ".");
-							}
 						}
 					}
 				}