+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/web/controller/CertificationIdController.java
@@ -7,7 +7,6 @@ import java.util.Map;
 import java.util.SortedSet;
 import java.util.TreeSet;
 
-import org.aspectj.lang.annotation.Aspect;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
 import org.springframework.security.access.annotation.Secured;
@@ -41,7 +40,6 @@ import io.swagger.annotations.ApiOperation;
 @Api(value="certification-ids")
 @RestController
 @RequestMapping("/certification_ids")
-@Aspect
 public class CertificationIdController {
 
 	@Autowired CertifiedProductManager certifiedProductManager;