+++ chpl/chpl-service/src/test/java/gov/healthit/chpl/dao/impl/CertificationStatusDAOTest.java
@@ -1,15 +1,10 @@
-package gov.healthit.chpl.web.controller;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.fail;
+package gov.healthit.chpl.dao.impl;
 
 import org.junit.BeforeClass;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.test.annotation.Rollback;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.TestExecutionListeners;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
@@ -24,8 +19,12 @@ import com.github.springtestdbunit.annotation.DatabaseSetup;
 import gov.healthit.chpl.auth.permission.GrantedPermission;
 import gov.healthit.chpl.auth.user.JWTAuthenticatedUser;
 import gov.healthit.chpl.caching.UnitTestRules;
+import gov.healthit.chpl.dao.CertificationStatusDAO;
+import gov.healthit.chpl.dao.EntityCreationException;
 import gov.healthit.chpl.dao.EntityRetrievalException;
-import gov.healthit.chpl.domain.Product;
+import gov.healthit.chpl.dto.CertificationStatusDTO;
+import gov.healthit.chpl.entity.CertificationStatusType;
+import junit.framework.TestCase;
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration(classes = { gov.healthit.chpl.CHPLTestConfig.class })
@@ -33,12 +32,13 @@ import gov.healthit.chpl.domain.Product;
     DirtiesContextTestExecutionListener.class,
     TransactionalTestExecutionListener.class,
     DbUnitTestExecutionListener.class })
-@DatabaseSetup("classpath:data/testData.xml") 
-public class ProductControllerTest {
-	@Autowired ProductController productController;
+@DatabaseSetup("classpath:data/testData.xml")
+public class CertificationStatusDAOTest extends TestCase {
 	@Rule
     @Autowired
     public UnitTestRules cacheInvalidationRule;
+	
+	@Autowired private CertificationStatusDAO certStatusDao;
 	
 	private static JWTAuthenticatedUser adminUser;
 	
@@ -52,22 +52,33 @@ public class ProductControllerTest {
 		adminUser.getPermissions().add(new GrantedPermission("ROLE_ADMIN"));
 	}
 	
+	/**
+	 * Tests that getByStatusName() completes without error for each status name
+	 * @throws EntityRetrievalException
+	 * @throws EntityCreationException
+	 */
 	@Test
 	@Transactional
-	@Rollback(true)
-	public void getProductById() {
-		Product result = null;
-		try {
-			result = productController.getProductById(-1L);
-		} catch(EntityRetrievalException e) {
-			fail(e.getMessage());
-		}
-		assertNotNull(result);
-		assertNotNull(result.getOwner());
-		assertNotNull(result.getOwner().getDeveloperId());
-		assertEquals(-1, result.getOwner().getDeveloperId().longValue());
-		assertNotNull(result.getOwnerHistory());
-		assertEquals(1, result.getOwnerHistory().size());
-		assertEquals(-2, result.getOwnerHistory().get(0).getDeveloper().getDeveloperId().longValue());
+	public void testgetByStatusName_Active() throws EntityRetrievalException, EntityCreationException {
+		System.out.println("Running testgetByStatusName() test");
+		CertificationStatusDTO dto = certStatusDao.getByStatusName(CertificationStatusType.Active.getName());
+		assertNotNull(dto);
+		dto = certStatusDao.getByStatusName(CertificationStatusType.Retired.getName());
+		assertNotNull(dto);
+		dto = certStatusDao.getByStatusName(CertificationStatusType.Pending.getName());
+		assertNotNull(dto);
+		dto = certStatusDao.getByStatusName(CertificationStatusType.WithdrawnByDeveloper.getName());
+		assertNotNull(dto);
+		dto = certStatusDao.getByStatusName(CertificationStatusType.WithdrawnByDeveloperUnderReview.getName());
+		assertNotNull(dto);
+		dto = certStatusDao.getByStatusName(CertificationStatusType.WithdrawnByAcb.getName());
+		assertNotNull(dto);
+		dto = certStatusDao.getByStatusName(CertificationStatusType.SuspendedByAcb.getName());
+		assertNotNull(dto);
+		dto = certStatusDao.getByStatusName(CertificationStatusType.SuspendedByOnc.getName());
+		assertNotNull(dto);
+		dto = certStatusDao.getByStatusName(CertificationStatusType.TerminatedByOnc.getName());
+		assertNotNull(dto);
 	}
+	
 }