+++ RELEASE_NOTES.md
@@ -1,5 +1,13 @@
 # Release Notes
 
+## Verison TBD
+_Date TBD_
+
+### Features Added
+* Add G1 and G2 macra measures to certified product upload, edit and detail display
+
+---
+
 ## Version TBD
 _Date TBD_
 
@@ -22,7 +30,7 @@ _Date TBD_
 * Add new certification status for products
 * Allow 0 for number of randomized sites and total sites for surveillance
 * Remove products marked Suspended By ONC from Decertified Product search
-* Add G1 and G2 macra measures to certified product upload, edit, and detail display
+* Allow retired test tools where certified product ICS=true
 
 ### Bugs Fixed
 * Fixes exception when getting back activity performed by a user that has since been deleted